requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
black==23.12.1
click==8.1.7
fastapi==0.109.0
greenlet==3.0.3
idna==3.6
Mako==1.3.0
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.0
typing_extensions==4.9.0
